+++ colors.py
@@ -1,0 +1,18 @@
+# referrence: http://htmlcolorcodes.com/
+
+C_black = (0, 0, 0)
+C_white = (255, 255, 255)
+C_red = (255, 0, 0)
+C_lime = (0, 255, 0)
+C_blue = (0, 0, 255)
+C_yellow = (255, 255, 0)
+C_aqua = (0, 255, 255)
+C_fuchsia = (255, 0, 255)
+C_silver = (192, 192, 192)
+C_gray = (128, 128, 128)
+C_maroon = (128, 0, 0)
+C_olive = (128, 128, 0)
+C_green = (0, 128, 0)
+C_purple = (128, 0, 128)
+C_teal = (0, 128, 128)
+C_navy = (0, 0, 128)