+++ stim/sound.py
@@ -1,6 +1,7 @@
 import math
 import numpy as np
 import scipy.io.wavfile
+
 from expy import shared
 from expy.response import *